Create a web server in python with basic auth
Created a web server in python which uses basic authentication. You can
deploy and reach the server at http://127.0.0.1:8000/ it will ask for
your credentials, if the credentials are correct it will re-direct you
to /token with a HTTP code 200 and will display a "Successful
Authentication" message. If the credentials are not correct it will not
redirect to /token and will display the message "Login failed"