Create Kubernetes manifests
I created two kubernete manifests one for the deployment and other for
the service to expose the deployment. The Deployment creates one replica
and uses the image created before. The label is defined as "app:
python-webserver". This label is used in the service to be able to
expose the desired deployment, the service is a LoadBalancer type which
means it will create a LoadBalancer in your cloud provider to expose the
application to the internet.